refactor the deviceConfiguration controller
